+++ engine/lite/kafka/integration-test/src/it/scala/pl/touk/nussknacker/engine/lite/kafka/NuReqRespRuntimeDockerTest.scala
@@ -29,4 +29,70 @@ class NuReqRespRuntimeDockerTest extends AnyFunSuite with ForAllTestContainer wi
     request.body(jsonPingMessage("dockerFoo")).send().body shouldBe Right(jsonPongMessage("dockerFoo"))
   }
 
+  test("should get scenario definition via http") {
+    val definitionReq = basicRequest.get(uri"http://${runtimeContainer.host}".port(mappedRuntimeApiPort).path("scenario", pingPongScenario.id, "definition"))
+
+    val definition = definitionReq.send().body.rightValue
+
+    definition shouldBe expectedOpenApiDef
+  }
+
+  private val expectedOpenApiDef =
+    s"""{
+       |  "openapi" : "3.1.0",
+       |  "info" : {
+       |    "title" : "${pingPongScenario.id}",
+       |    "version" : "1"
+       |  },
+       |  "paths" : {
+       |    "/reqresp-ping-pong" : {
+       |      "post" : {
+       |        "description" : "**scenario name**: reqresp-ping-pong",
+       |        "tags" : [
+       |          "Nussknacker"
+       |        ],
+       |        "requestBody" : {
+       |          "required" : true,
+       |          "content" : {
+       |            "application/json" : {
+       |              "schema" : {
+       |                "type" : "object",
+       |                "nullable" : false,
+       |                "properties" : {
+       |                  "ping" : {
+       |                    "type" : "string",
+       |                    "nullable" : false
+       |                  }
+       |                }
+       |              }
+       |            }
+       |          }
+       |        },
+       |        "produces" : [
+       |          "application/json"
+       |        ],
+       |        "consumes" : [
+       |          "application/json"
+       |        ],
+       |        "summary" : "reqresp-ping-pong",
+       |        "responses" : {
+       |          "200" : {
+       |            "content" : {
+       |              "application/json" : {
+       |                "schema" : {
+       |                  "type" : "object",
+       |                  "properties" : {
+       |                    "pong" : {
+       |                      "type" : "string"
+       |                    }
+       |                  }
+       |                }
+       |              }
+       |            }
+       |          }
+       |        }
+       |      }
+       |    }
+       |  }
+       |}""".stripMargin
 }